+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 0.2.1 (unreleased)
 
+* Confirm before destroying server
+* Handle BoxURL before validate on box startup.
+
 # 0.2.0 (April 19, 2013)
 
 * Merge from upstream 'vagrant-aws' v0.2.2